third_party/blink/web_tests/platform/generic/svg/custom/svg-xml-dom-sync-expected.txt

Tests whether changes to SVG through native objects and the DOM stay in sync.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rect.x.baseVal.value is 100
PASS rect.getAttribute('x') is '100'
PASS rect.y.baseVal.value is 100
PASS rect.getAttribute('y') is '100'
PASS rect.x.baseVal.value is 200
PASS rect.getAttribute('x') is '200'
PASS rect.y.baseVal.value is 200
PASS rect.getAttribute('y') is '200'
PASS successfullyParsed is true

TEST COMPLETE
